backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pdfplumber==0.11.4
pypdf==5.1.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

